+++ server/output/chapter01.h
@@ -3,13 +3,6 @@
 
 #include <loke/scene.h>
 
-// Forward declarations for all scenes in this chapter
-void scene_forrest(GameState* state);
-void scene_hills(GameState* state);
-void scene_house(GameState* state);
-void scene_path(GameState* state);
-void scene_playwright_test(GameState* state);
-void scene_test_codegen(GameState* state);
-void scene_township(GameState* state);
+// No scenes yet
 
 #endif // CHAPTER01_H